Working around Swift 4 issue on linux where linux `JSONSerialization` returns a different type for numeric arrays when passed 1.0.
```
var coordinates1 = try JSONSerialization.jsonObject(with: Data(bytes: Array("[1.0, 1.0]".utf8))) as? [Any]

type(of: coordinates1![0])
```

The above code returns `Int` on linux and `NSNumber` on OSX/iOS.  This causes our elegant cast conversion to fail on the current snapshot of Swift 4 (swift-4.0-DEVELOPMENT-SNAPSHOT-2017-08-04-a-ubuntu16.10) because it can't cast the `Int` to a `Double` as required.  This fix works around the issue by converting to an `Any` type of array first and then testing the type of each element and converting it manually.